Added interface and reorganized classes
Added buffer interface and reorganized classes.
