+++ src/main/java/code/elix_x/excore/utils/client/render/IVertexBuffer.java
@@ -1,0 +1,9 @@
+package code.elix_x.excore.utils.client.render;
+
+public interface IVertexBuffer {
+
+	void draw();
+
+	void cleanUp();
+
+}
