+++ src/main/java/code/elix_x/excore/utils/client/render/vao/VertexBufferSeparate.java
@@ -1,10 +1,11 @@
-package code.elix_x.excore.utils.client.render;
+package code.elix_x.excore.utils.client.render.vao;
 
 import java.nio.ByteBuffer;
 
 import org.lwjgl.opengl.GL15;
 import org.lwjgl.opengl.GL20;
 
+import code.elix_x.excore.utils.client.render.vbo.VBO;
 import net.minecraft.client.renderer.vertex.VertexFormat;
 import net.minecraft.client.renderer.vertex.VertexFormatElement;
 import net.minecraft.client.renderer.vertex.VertexFormatElement.EnumUsage;
