+++ src/main/java/code/elix_x/excore/utils/client/render/vbo/VBO.java
@@ -1,4 +1,4 @@
-package code.elix_x.excore.utils.client.render;
+package code.elix_x.excore.utils.client.render.vbo;
 
 import java.nio.ByteBuffer;
